### 2024-11-08 dac10c3
Update paper.bib
add dois
