+++ docs/paper/paper.bib
@@ -8,7 +8,8 @@
   pages     = "1427--1437",
   month     =  mar,
   year      =  2012,
-  language  = "en"
+  language  = "en",
+  doi = "10.1007/s12665-011-1146-1"
 }
 
 @ARTICLE{Chahinian2019-lg,
@@ -24,7 +25,8 @@
   pages     = "101370",
   month     =  nov,
   year      =  2019,
-  language  = "en"
+  language  = "en",
+  doi = "10.1016/j.compenvurbsys.2019.101370",
 }
 
 % The entry below contains non-ASCII chars that could not be converted
@@ -72,7 +74,8 @@
   month     =  jan,
   year      =  2022,
   copyright = "http://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi = "10.1029/2021wr030778"
 }
 <!-- markdownlint-disable MD009 -->
 @ARTICLE{Rauch2017-jz,
@@ -109,7 +112,8 @@
   pages     = "501--514",
   month     =  dec,
   year      =  2017,
-  language  = "en"
+  language  = "en",
+  doi = "10.1016/j.watres.2017.09.039"
 }
 <!---  --->
 
@@ -124,7 +128,8 @@
   pages     = "517--533",
   month     =  jan,
   year      =  2019,
-  language  = "en"
+  language  = "en",
+  doi = "10.1016/j.jhydrol.2018.11.005"
 }
 
 @ARTICLE{Chegini2022-oo,
@@ -160,7 +165,8 @@
   month     =  aug,
   year      =  2022,
   copyright = "https://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi = "10.5194/hess-26-4279-2022",
 }
 
 @ARTICLE{Reyes-Silva2022-pr,
@@ -196,7 +202,8 @@
   month     =  dec,
   year      =  2022,
   copyright = "https://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi = "10.3390/w15010046"
 }
 
 @UNPUBLISHED{Dobson2024-dv,
@@ -229,7 +236,8 @@
               where traditional approaches are impractical.",
   journal  = "EarthArXiv",
   month    =  oct,
-  year     =  2024
+  year     =  2024,
+  doi = "10.31223/X5GT5X"
 }
 
 @ARTICLE{Khurelbaatar2021-sp,
@@ -283,7 +291,8 @@
   volume    =  9,
   month     =  may,
   year      =  2021,
-  copyright = "https://creativecommons.org/licenses/by/4.0/"
+  copyright = "https://creativecommons.org/licenses/by/4.0/",
+  doi = "10.3389/fenvs.2021.626634"
 }
 
 @misc{VIDA2023,
