### 2024-11-08 #7438 @barneydobson
Thanks @kthyng  - TIL mendeley bibtex export doesn't include DOI 🤦‍♂️
