+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
       - id: check-merge-conflict
       - id: debug-statements
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.7.3"
+    rev: "v0.7.4"
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix]
