+++ .github/workflows/ci_template.yml
@@ -37,7 +37,7 @@ jobs:
 
     - name: Upload coverage to Codecov
       if: success() && (matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10')
-      uses: codecov/codecov-action@v4
+      uses: codecov/codecov-action@v5
       with:
         fail_ci_if_error: true
         token: ${{ secrets.codecov_token }}
