### 2024-12-09 2a2a779
Update ci_template.yml
think that's all I need...
