+++ .github/workflows/ci_template.yml
@@ -44,13 +44,13 @@ jobs:
         verbose: true
 
   extended_demo:
-    runs-on: macos-latest
+    runs-on: macos-12
     steps:
     - uses: actions/checkout@v4
     
     - uses: actions/setup-python@v5
       with:
-        python-version: '3.10'
+        python-version: '3.10.15'
 
     - name: Install dependencies
       run: pip install -e .[dev,doc]
