### 2024-12-12 9918113
Update design_graphfcns.py
