### 2024-12-12 6f05fc3
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
