### 2024-12-12 9fc06e5
Update design_graphfcns.py
