+++ src/swmmanywhere/graphfcns/design_graphfcns.py
@@ -157,7 +157,7 @@ def process_successors(
     tot = sum([G.nodes[anc_node]["contributing_area"] for anc_node in anc])
     M3_PER_HR_TO_M3_PER_S = 1 / 60 / 60
     Q = tot * hydraulic_design.precipitation * M3_PER_HR_TO_M3_PER_S
-    
+
     for ix, ds_node in enumerate(G.successors(node)):
         edge = G.get_edge_data(node, ds_node, 0)
 
