### 2025-01-27 #7729 @editorialbot
[REVIEW]: SWMManywhere: Synthesise Urban Drainage Network Models Anywhere in the World
**Submitting author:** @barneydobson (<a href="http://orcid.org/0000-0002-0149-4124">Barnaby Dobson</a>)
**Repository:** https://github.com/ImperialCollegeLondon/SWMManywhere
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.16
