### 2025-02-21 d5b4e48
Update CITATION.cff
