### 2025-02-21 0196b4d
Update CITATION.cff
