+++ CITATION.cff
@@ -5,7 +5,7 @@ authors:
     given-names: "Barnaby"
     orcid: "https://orcid.org/0000-0002-0149-4124"
   - family-names: "Jovanovic"
-    orcid: ""
+    given-names: "Tijana"
   - family-names: "Alonso-Álvarez"
     given-names: "Diego"
     orcid: "https://orcid.org/0000-0002-0060-9495"
