### 2025-02-21 7a1a4e4
Update CITATION.cff
