### 2025-02-21 750fe4a
Update CITATION.cff
