### 2025-03-02 ImperialCollegeLondon/SWMManywhere#364 @meghnathomas
JOSS review
Hi! I am opening this issue to hold all my comments for your JOSS review. SWMManywhere is a cool concept and could prove to be a useful tool to
build urban drainage network models from scratch, and I like the graph functions and metrics capabilities. I will add comments to this thread as I continue to explore the package, but for now here is an initial round of comments and general questions. Feel free to refute any of my suggestions, and go ahead and tick the boxes if you have addressed these comments!

General

Docs (still in progress)

- [ ] 4. I like the detailed commentary in the "Extended Demo" page very much!

Paper (still in progress)
