### 2025-03-03 ImperialCollegeLondon/SWMManywhere#364 @barneydobson
Thanks so much for your detailed comments @meghnathomas . I will make each of these a separate issue to ensure the discussion doesn't get overcomplicated and reply in those issues.
