+++ docs/paper/paper.bib
@@ -206,38 +206,13 @@
   doi = "10.3390/w15010046"
 }
 
-@UNPUBLISHED{Dobson2024-dv,
-  title    = "{SWMManywhere}: A global-scale workflow for generation and
-              sensitivity analysis of synthetic urban drainage models",
-  author   = "Dobson, Barnaby and Jovanovic, Tijana and Alonso-{\'A}lvarez,
-              Diego and Chegini, Taher",
-  abstract = "Continual improvements in publicly available global geospatial
-              datasets provide an opportunity for deriving urban drainage
-              networks and simulation models of these networks (UDMs)
-              worldwide. We present SWMManywhere, which leverages such datasets
-              for generating synthetic UDMs and creating a Storm Water
-              Management Model for any urban area globally. SWMManywhere's
-              highly modular and parameterised approach enables significant
-              customisation to explore hydraulicly feasible network
-              configurations. Key novelties of our workflow are in network
-              topology derivation that accounts for combined effects of
-              impervious area and pipe slope. We assess SWMManywhere by
-              comparing pluvial flooding, drainage network outflows, and design
-              with known networks. The results demonstrate high quality
-              simulations are achievable with a synthetic approach even for
-              large networks. Our extensive sensitivity analysis shows that the
-              locations of manholes, outfalls, and underlying street network
-              are the most sensitive parameters. We find widespread sensitivity
-              across all parameters without clearly defined values that they
-              should take, thus, recommending an uncertainty driven approach to
-              synthetic drainage network modelling. This study showcases
-              significant potential of SWMManywhere for research and industry
-              applications to provide drainage network models in urban areas
-              where traditional approaches are impractical.",
-  journal  = "EarthArXiv",
-  month    =  oct,
-  year     =  2024,
-  doi = "10.31223/X5GT5X"
+@article{Dobson2025,
+author = {Dobson, Barnaby and Jovanovic, Tijana and Alonso-Álvarez, Diego and Chegini, Taher},
+doi = {10.1016/j.envsoft.2025.106358},
+journal = {Environmental Modelling \& Software},
+title = {{SWMManywhere: A Workflow for Generation and Sensitivity Analysis of Synthetic Urban Drainage Models, Anywhere}},
+url = {https://doi.org/10.1016/j.envsoft.2025.106358},
+year = {2025}
 }
 
 @ARTICLE{Khurelbaatar2021-sp,
