### 2025-03-03 9c2744e
Update paper.md
add mroe description
