+++ docs/paper/paper.bib
@@ -318,3 +318,14 @@ year = {2017}
    howpublished = "\url{ https://overturemaps.org/ }",
    year = {2024},
  }
+
+ @article{sanne2024pysewer,
+  title={Pysewer: A Python Library for Sewer Network Generation in Data Scarce Regions},
+  author={Sanne, Moritz and Khurelbaatar, Ganbaatar and Despot, Daneish and van Afferden, Manfred and Friesen, Jan},
+  journal={Journal of Open Source Software},
+  volume={9},
+  number={104},
+  pages={6430},
+  year={2024},
+  doi={10.21105/joss.06430},
+}
