+++ docs/paper/paper.bib
@@ -329,3 +329,14 @@ year = {2017}
   year={2024},
   doi={10.21105/joss.06430},
 }
+
+@article{mcdonnell2020pyswmm,
+  title={PySWMM: the python interface to stormwater management model (SWMM)},
+  author={McDonnell, Bryant E and Ratliff, Katherine and Tryby, Michael E and Wu, Jennifer Jia Xin and Mullapudi, Abhiram},
+  journal={Journal of open source software},
+  volume={5},
+  number={52},
+  pages={1},
+  year={2020},
+  doi={10.21105/joss.02292}
+}
