### 2025-03-03 f57b442
Update paper.md
explain difference between EMS and JOSS paper
