### 2025-03-03 2a2c673
Update test_graph_utilities.py
fix test
