### 2025-03-05 26ec69c
add run_model flag
