### 2025-03-05 e823d5b
allow specification of modules from config
