### 2025-03-06 9bff49f
Update parameters_guide.md
finish doc
