### 2025-03-10 ImperialCollegeLondon/SWMManywhere#384 @cbuahin
JOSS Review
This issue has been created to provide comments as part of my review of the SWMManywhere software package for the  Journal of Open Source Software (openjournals/joss-reviews#7729).

The SWMManywhere package is a worthwhile effort that attempts reduce the burden of model domain delineation and parameterization for the USEPA Stormwater Management Model (SWMM) by utilizing publicly and freely available geospatial datasets including street networks, digital elevation models, building footprints, etc.

I will be adding comments, questions, and general suggestions to this thread as I proceed with my review. Feel free to check the boxes for those items you have addressed below.

# General Comments

- [ ]
