+++ .github/workflows/ci_template.yml
@@ -12,7 +12,7 @@ on:
         type: string
       python-version:
         description: 'Python version'
-        default: '["3.10"]'
+        default: '["3.11"]'
         type: string
 
 jobs:
