+++ pyproject.toml
@@ -60,6 +60,7 @@ dependencies = [
   "rioxarray",
   "scipy",
   "shapely",
+  "swmmio",
   "tqdm",
   "xarray",
 ]
