### 2025-03-31 47a316e
allow plot_map for geojson
