+++ src/swmmanywhere/utilities.py
@@ -57,48 +57,29 @@ def yaml_dump(o: Any, stream: Any = None, **kwargs: Any) -> str:
     )
 
 
-def load_edges_nodes(model_dir: Path) -> tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]:
-    """Load edges and nodes from a model directory.
-
-    Args:
-        model_dir (Path): The directory containing the model files.
-
-        Returns:
-            tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]: A tuple containing the nodes and edges GeoDataFrames.
-    """
-    if (model_dir / "nodes.geoparquet").exists():
-        nodes = gpd.read_parquet(model_dir / "nodes.geoparquet")
-    elif (model_dir / "nodes.geojson").exists():
-        nodes = gpd.read_file(model_dir / "nodes.geojson")
-    else:
-        raise FileNotFoundError("No nodes found in model directory.")
-    if (model_dir / "edges.geoparquet").exists():
-        edges = gpd.read_file(model_dir / "edges.geoparquet")
-    elif (model_dir / "edges.geojson").exists():
-        edges = gpd.read_file(model_dir / "edges.geojson")
-    else:
-        raise FileNotFoundError("No edges found in model directory.")
-    
-    return nodes, edges
-
-def plot_basic(model_dir: Path):
+def plot_basic(nodes_path: Path, edges_path: Path):
     """Create a basic map with nodes and edges.
 
     Args:
-        model_dir (Path): The directory containing the model files.
+        nodes_path (Path): The path to the nodes file.
+        edges_path (Path): The path to the edges file.
 
     Returns:
         folium.Map: The folium map.
     """
     # Load and inspect results
-    nodes, edges = load_edges_nodes(model_dir)
+    nodes = gpd.read_file(nodes_path)
+    edges = gpd.read_file(edges_path)
 
     # Convert to EPSG 4326 for plotting
     nodes = nodes.to_crs(4326)
     edges = edges.to_crs(4326)
 
     # Identify outfalls
-    outfall = nodes.id == nodes.outfall
+    if "outfall" in nodes.columns:
+        outfall = nodes.id == nodes.outfall
+    else:
+        outfall = nodes.id == None
 
     # Plot on map
     m = folium.Map(
@@ -129,18 +110,21 @@ def plot_basic(model_dir: Path):
     return m
 
 
-def plot_clickable(model_dir: Path):
+def plot_clickable(nodes_path: Path, edges_path: Path, results_path: Path):
     """Create a clickable map with nodes, edges and results.
 
     Args:
-        model_dir (Path): The directory containing the model files.
+        nodes_path (Path): The path to the nodes file.
+        edges_path (Path): The path to the edges file.
+        results_path (Path): The path to the results file.
 
     Returns:
         folium.Map: The folium map.
     """
     # Load and inspect results
-    nodes, edges = load_edges_nodes(model_dir)
-    df = pd.read_parquet(model_dir / "results.parquet")
+    nodes = gpd.read_file(nodes_path)
+    edges = gpd.read_file(edges_path)
+    df = pd.read_parquet(results_path)
     df.id = df.id.astype(str)
     floods = df.loc[df.variable == "flooding"].groupby("id")
     flows = df.loc[df.variable == "flow"].groupby("id")
@@ -190,7 +174,7 @@ def plot_clickable(model_dir: Path):
         img.seek(0)
         img_base64 = base64.b64encode(img.read()).decode()
         img_html = f'<img src="data:image/png;base64,{img_base64}">'
-        if row.outfall == node:
+        if row.get("outfall") == node:
             color = "red"
         else:
             color = "black"
[…]
