### 2025-04-01 #7729 @cbuahin
@mengqi-z  and @barneydobson , I have provided some initial questions and suggestions for the authors' consideration. I am doing a deeper dive on the package codebase and may provide some additional questions/suggestions.
