### 2025-04-01 ImperialCollegeLondon/SWMManywhere#384 @barneydobson
@cbuahin - thanks for the comments there, I have converted them to sub-issues to be addressed separately. Did you want me to add you as a collaborator so that you can review the PRs that fix them? Or do you want me to just let you know here when a PR has been closed so that you can check it?
