### 2025-04-01 ImperialCollegeLondon/SWMManywhere#384 @cbuahin
@barneydobson , go ahead and add me as a reviewer so that I am notified. On a separate note, I maintain EPA'S SWMM software and we are developing an official python binding for SWMM that we intend to maintain as we embark on advancing the SWMM engine. Consider checking it out as it may be useful to integrate it into your efforts in the future.

https://pypi.org/project/epaswmm/
