### 2025-04-01 ImperialCollegeLondon/SWMManywhere#384 @barneydobson
@cbuahin OK added for triage, I guess it's fine but happy to change as needed.

That's great that there could be more expansive/official python bindings! I've been using `pyswmm` well enough to date but welcome any further control over the simulations that could be achieved.
