+++ src/swmmanywhere/utilities.py
@@ -79,7 +79,9 @@ def plot_basic(nodes_path: Path, edges_path: Path):
     if "outfall" in nodes.columns:
         outfall = nodes.id == nodes.outfall
     else:
-        outfall = nodes.id == None
+        outfall = nodes.id.isin(edges.u.astype(str)) & ~nodes.id.isin(
+            edges.v.astype(str)
+        )
 
     # Plot on map
     m = folium.Map(
