+++ src/swmmanywhere/defs/__init__.py
@@ -1,0 +1,5 @@
+"""Module for data shipped with the package."""
+
+from __future__ import annotations
+
+from .copy_test_data import copy_test_data
