### 2025-04-01 17a3c99
little refactor
