+++ pyproject.toml
@@ -60,7 +60,6 @@ dependencies = [
   "rioxarray",
   "scipy",
   "shapely",
-  "swmmio",
   "tqdm",
   "xarray",
 ]
