### 2025-04-02 c3619c7
Create custom_data_demo.py
initial commit
