+++ docs/notebooks/custom_data_demo.py
@@ -1,7 +1,7 @@
 # %% [markdown]
 # # Custom data demo
 # Note - this script can also be opened in interactive Python if you wanted to
-# play around. On the GitHub it is in [docs/notebooks](https://github.com/ImperialCollegeLondon/SWMManywhere/blob/main/docs/notebooks/extended_demo.py)
+# play around. On the GitHub it is in [docs/notebooks](./notebooks/extended_demo.py)
 # . To run this on your
 # local machine, you will need to install the optional dependencies for `doc`:
 #
@@ -11,19 +11,22 @@
 # ## Introduction
 # This script demonstrates how to use `swmmanywhere` when you have custom data.
 #
-# Since this is a notebook, we will define [`config`](https://imperialcollegelondon.github.io/SWMManywhere/config_guide/)
+# Since this is a notebook, we will define [`config`](config_guide.md)
 # as a dictionary rather than a `yaml` file, but the same principles apply.
 #
 # ## Initial setup
 #
-# We will use the same example as the [extended demo](extended_demo.py), but with a
-# custom elevation dataset. Let's start by rerunning it.
+# We will use the same example as the [extended demo](./notebooks/extended_demo.py), but
+# with a custom elevation dataset. Let's start by rerunning it.
 # %%
 # Imports
 from __future__ import annotations
 
 import tempfile
 from pathlib import Path
+
+import folium
+import geopandas as gpd
 
 from swmmanywhere.logging import set_verbose
 from swmmanywhere.swmmanywhere import swmmanywhere
@@ -57,8 +60,17 @@ set_verbose(True)  # Set verbosity
 outputs = swmmanywhere(config)
 model_dir = outputs[0].parent
 
+# %% [markdown]
+# ## Plotting output
+#
+# Now we can plot the output. To highlight the differences in the supplied data that we
+# are about to demonstrate, we also plot the subbasins.
 # %%
-plot_map(model_dir)
+m = plot_map(model_dir)
+subbasins = gpd.read_file(model_dir / "subbasins.geoparquet")
+folium.GeoJson(subbasins, fill_opacity=0, color="blue", weight=2).add_to(m)
+m
+
 
 # %% [markdown]
 # ## Supply custom elevation data
@@ -108,5 +120,28 @@ config["address_overrides"] = {
 outputs = swmmanywhere(config)
 model_dir = outputs[0].parent
 
+# %% [markdown]
+# ## Plotting output
+#
+# This time we will include both the original (blue) and the new (red) subbasins to
+# highlight the impact of flipping the elevation data.
 # %%
-plot_map(model_dir)
+m = plot_map(model_dir)
+subbasins_new = gpd.read_file(model_dir / "subbasins.geoparquet")
+folium.GeoJson(subbasins_new, fill_opacity=0, color="red", weight=2).add_to(m)
+folium.GeoJson(subbasins, fill_opacity=0, color="blue", weight=2).add_to(m)
+m
+
+# %% [markdown]
+#
+# ## Adding data to SWMManywhere
+#
+# Now we can see how to add our own data to SWMManywhere. This method works for any
+# [filepath](reference-filepaths.md), which are the files generated by the current
+# workflow if verbosity is true (i.e., `swmmanywhere.logging.set_verbose(True)`).
+#
+# You can also use `address_overrides` to you add your own custom data that
+# might be used by a [custom graphfcn](graphfcns_guide.md), for example. If you are
+# adding new data to the SWMManywhere workflow in this way, you will have to access it
+# from the `FilePaths` object with the `get_path` method.
+#
