+++ docs/notebooks/custom_data_demo.py
@@ -1,9 +1,9 @@
 # %% [markdown]
 # # Custom data demo
 # Note - this script can also be opened in interactive Python if you wanted to
-# play around. On the GitHub it is in [docs/notebooks](./notebooks/extended_demo.py)
-# . To run this on your
-# local machine, you will need to install the optional dependencies for `doc`:
+# play around. On the GitHub it is in [docs/notebooks](./notebooks/custom_data_demo.py).
+# To run this on your local machine, you will need to install the optional dependencies
+# for `doc`:
 #
 # `pip install swmmanywhere[doc]`
 #
@@ -11,7 +11,7 @@
 # ## Introduction
 # This script demonstrates how to use `swmmanywhere` when you have custom data.
 #
-# Since this is a notebook, we will define [`config`](config_guide.md)
+# Since this is a notebook, we will define [`config`](https://imperialcollegelondon.github.io/SWMManywhere/config_guide/)
 # as a dictionary rather than a `yaml` file, but the same principles apply.
 #
 # ## Initial setup
@@ -79,6 +79,8 @@ m
 # and perturb it, though in practice you are likely to use some higher resolution
 # or more accurate data.
 #
+# You don't need to worry about your files lining up perfectly (though of course if they
+# do not overlap at all then you will run into problems).
 # %%
 
 # Import NASADEM downloader and reprojection tools
@@ -137,11 +139,11 @@ m
 # ## Adding data to SWMManywhere
 #
 # Now we can see how to add our own data to SWMManywhere. This method works for any
-# [filepath](reference-filepaths.md), which are the files generated by the current
-# workflow if verbosity is true (i.e., `swmmanywhere.logging.set_verbose(True)`).
+# [filepath](https://imperialcollegelondon.github.io/SWMManywhere/reference-filepaths/),
+# which are the files used by the current workflow.
 #
 # You can also use `address_overrides` to you add your own custom data that
-# might be used by a [custom graphfcn](graphfcns_guide.md), for example. If you are
-# adding new data to the SWMManywhere workflow in this way, you will have to access it
-# from the `FilePaths` object with the `get_path` method.
+# might be used by a [custom graphfcn](https://imperialcollegelondon.github.io/SWMManywhere/graphfcns_guide/)
+# , for example. If you are adding new data to the SWMManywhere workflow in this way,
+# you will have to access it from the `FilePaths` object with the `get_path` method.
 #
