### 2025-04-02 0699787
Update tests/test_swmmanywhere.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
