### 2025-04-03 b71a428
fix, track and test costs
