### 2025-04-03 301e296
add zip stuff
