+++ .github/workflows/ci_template.yml
@@ -31,6 +31,11 @@ jobs:
 
     - name: Install dependencies
       run: pip install -e .[dev]
+    
+    - name: Checkout repo with submodules
+      uses: actions/checkout@v4
+      with:
+        submodules: recursive
 
     - name: Run tests
       run: pytest
