+++ .github/workflows/ci_template.yml
@@ -31,14 +31,9 @@ jobs:
 
     - name: Install dependencies
       run: pip install -e .[dev]
-    
-    - name: Checkout repo with submodules
-      uses: actions/checkout@v4
-      with:
-        submodules: recursive
 
     - name: Run tests
-      run: pytest --ignore=tests/pywbt_source/tests
+      run: pytest
 
     - name: Upload coverage to Codecov
       if: success() && (matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10')
