+++ tests/conftest.py
@@ -14,7 +14,10 @@ def pytest_collection_modifyitems(config, items):
 
 @pytest.fixture
 def wbt_path() -> str:
-    """Determine the platform suffix for downloading WhiteboxTools."""
+    """Determine the platform suffix for downloading WhiteboxTools.
+    
+    Based on implementation in [`pywbt`](https://github.com/cheginit/pywbt).
+    """
     system = platform.system()
     base_name = "WhiteboxTools_{}.zip"
     if system not in ("Windows", "Darwin", "Linux"):
