### 2025-04-03 76c1d51
Update src/swmmanywhere/defs/copy_test_data.py
Co-authored-by: Diego Alonso Álvarez <6095790+dalonsoa@users.noreply.github.com>
