+++ src/swmmanywhere/geospatial_utilities.py
@@ -24,7 +24,7 @@ import pyproj
 import rasterio as rst
 import rioxarray
 import shapely
-from pywbt import whitebox_tools
+from pywbt import whitebox_tools, prepare_wbt
 from rasterio import features
 from scipy.interpolate import RegularGridInterpolator
 from scipy.spatial import KDTree
@@ -565,7 +565,7 @@ def derive_subbasins_streamorder(
     streamorder: int | None = None,
     x: list[float] = [],
     y: list[float] = [],
-    wbt_path: Path | None = None,
+    wbt_zip_path: Path | None = None,
 ) -> gpd.GeoDataFrame:
     """Derive subbasins.
 
@@ -578,13 +578,13 @@ def derive_subbasins_streamorder(
         streamorder (int): The stream order to delineate subbasins for.
         x (list): X coordinates.
         y (list): Y coordinates.
-        wbt_path (Path, optional): Path to WhiteboxTools binaries. Defaults to None.
+        wbt_zip_path (Path, optional): Path to WhiteboxTools binaries. Defaults to None.
 
     Returns:
         gpd.GeoDataFrame: A GeoDataFrame containing polygons.
     """
     # Load and process the DEM
-    grid, flow_dir, _ = load_and_process_dem(fid, wbt_path=wbt_path)
+    grid, flow_dir, _ = load_and_process_dem(fid, wbt_zip_path=wbt_zip_path)
 
     flw = pyflwdir.from_array(
         flow_dir,
@@ -625,12 +625,12 @@ def derive_subbasins_streamorder(
     return gdf_bas
 
 
-def flwdir_whitebox(fid: Path, wbt_path: Path | None = None) -> np.array:
+def flwdir_whitebox(fid: Path, wbt_zip_path: Path | None = None) -> np.array:
     """Calculate flow direction using WhiteboxTools.
 
     Args:
         fid (Path): Filepath to the DEM.
-        wbt_path (Path, optional): Path to WhiteboxTools binaries. If the binaries are
+        wbt_zip_path (Path, optional): Path to WhiteboxTools binaries. If the binaries are
             not present, they will be downloaded. If None, the binaries will be stored
             temporarily and removed after the function call. Defaults to None.
 
@@ -646,25 +646,31 @@ def flwdir_whitebox(fid: Path, wbt_path: Path | None = None) -> np.array:
         dem = temp_path / "dem.tif"
         shutil.copy(fid, dem)
 
-        if not wbt_path:
-            wbt_path = fid.parent / "whiteboxtools_binaries.zip"
+        wbt_root = temp_path / "WBT"
+
+        if wbt_zip_path:
+            prepare_wbt(wbt_root, wbt_zip_path, refresh_download=False)
+        else:
+            prepare_wbt(wbt_root)
 
         # Condition
         wbt_args = {
             "BreachDepressions": ["-i=dem.tif", "--fillpits", "-o=dem_corr.tif"],
             "D8Pointer": ["-i=dem_corr.tif", "-o=fdir.tif"],
         }
-        wbt_root = temp_path / "WBT"
+        
         whitebox_tools(
             temp_path,
             wbt_args,
             save_dir=temp_path,
             verbose=verbose(),
             wbt_root=wbt_root,
-            zip_path=wbt_path,
+            zip_path=wbt_zip_path,
             max_procs=1,
         )
-        (wbt_root / "whitebox_tools.exe").unlink(missing_ok=True)
+
+        shutil.rmtree(wbt_root)
+
         fdir = temp_path / "fdir.tif"
         if not Path(fdir).exists():
             raise ValueError("Flow direction raster not created.")
[…]
