+++ tests/conftest.py
@@ -4,7 +4,7 @@ import platform
 from pathlib import Path
 
 import pytest
-
+from pywbt.pywbt import _get_platform_suffix
 
 def pytest_collection_modifyitems(config, items):
     """Skip tests marked with downloads."""
@@ -13,24 +13,15 @@ def pytest_collection_modifyitems(config, items):
 
 
 @pytest.fixture
-def wbt_path() -> Path:
+def wbt_zip_path() -> Path:
     """Determine the platform specific binary for whiteboxtools for testing.
 
     All WBT binaries are stored in `wbt_zip` within the `tests` directory.
 
     Based on implementation in [`pywbt`](https://github.com/cheginit/pywbt).
     """
-    system = platform.system()
     base_name = "WhiteboxTools_{}.zip"
-    if system not in ("Windows", "Darwin", "Linux"):
-        raise ValueError(f"Unsupported operating system: {system}")
+    
+    _, suffix, _ = _get_platform_suffix()
 
-    if system == "Windows":
-        suffix = "win_amd64"
-    elif system == "Darwin":
-        suffix = "darwin_m_series" if platform.machine() == "arm64" else "darwin_amd64"
-    else:
-        suffix = (
-            "linux_musl" if "musl" in platform.libc_ver()[0].lower() else "linux_amd64"
-        )
     return Path(__file__).parent / "wbt_zip" / base_name.format(suffix)
