+++ tests/conftest.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
-import platform
 from pathlib import Path
 
 import pytest
 from pywbt.pywbt import _get_platform_suffix
+
 
 def pytest_collection_modifyitems(config, items):
     """Skip tests marked with downloads."""
@@ -21,7 +21,7 @@ def wbt_zip_path() -> Path:
     Based on implementation in [`pywbt`](https://github.com/cheginit/pywbt).
     """
     base_name = "WhiteboxTools_{}.zip"
-    
+
     _, suffix, _ = _get_platform_suffix()
 
     return Path(__file__).parent / "wbt_zip" / base_name.format(suffix)
