+++ src/swmmanywhere/graphfcns/subcatchment_graphfcns.py
@@ -234,7 +234,9 @@ class calculate_contributing_area(
 
             # Derive
             subs_gdf = go.derive_subcatchments(
-                G, temp_fid, wbt_zip_path=addresses.project_paths.whiteboxtools_binaries_zip
+                G,
+                temp_fid,
+                wbt_zip_path=addresses.project_paths.whiteboxtools_binaries_zip,
             )
             if verbose():
                 subs_gdf.to_file(addresses.model_paths.subcatchments, driver="GeoJSON")
