### 2025-04-14 642b5a2
Update geospatial_utilities.py
