### 2025-04-14 d22dcfc
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.11.0 → v0.11.5](https://github.com/astral-sh/ruff-pre-commit/compare/v0.11.0...v0.11.5)
