### 2025-04-25 b6e57b1
add custom slope points
