### 2025-04-25 55ab9ca
Update test_graph_utilities.py
typo
