### 2025-04-28 ImperialCollegeLondon/SWMManywhere#364 @meghnathomas
Hi @barneydobson, thanks for your patience! I closed the issues relating to my comments that (in my opinion) you've addressed, but feel free to reopen them if you need. Here are some thoughts and comments:

1. The delay in my response has been because I've been struggling to test the various .py demos! I have tried to uninstall + reinstall swmmanywhere a few different ways, and my latest install has been by git cloning the user-data-example branch, locally installing swmmanywhere on my machine (navigating to the swmmanywhere folder and entering "pip install -e .") and then proceeding to run the extended demo, validation demo, and custom data demo files using Spyder, my preferred IDE. Do you have recommendations for what I should be doing differently? I am not able to run any of the three files to completion, and would be happy to share the error logs but I think it's too much information and I'm probably doing something wrong on my end.

2. In any case, I suggest that you merge the user-data-example and real-network-demo PRs since your team has approved those and I assume they work well on your end. From what I can see, the PRs satisfy my request for more fleshed out and applicable examples and I appreciate the effort you took to create them! I am happy to try again to install and run the code from a clean slate once everything has been merged and the PyPI version has been bumped.

3. Were you planning to add more information to the paper regarding #373 (or have you already done that and did I miss it)?

4. Once the PRs have been merged and I can run the code successfully, I think everything will be good to go on my side!
