+++ docs/paper/paper.md
@@ -64,7 +64,7 @@ We visualise the example from the [extended demonstration](https://imperialcolle
 
 ### Comparisons with real networks
 
-When comparing synthesised UDMs of an actual location with a known UDM, it is not trivial to evaluate the efficacy of the synthesis algorithm because pipes do not line up [@Chahinian2019-lg]. Thus, SWMManywhere contains an extensive evaluation framework that contains a high number of metrics for quantifying comparisons. Because each graph function in SWMManywhere depends on the output of the previous, we evaluate different types of metrics that span the synthesis process (broadly categorised to systemic, topographical, and design properties) in addition to the more conventional simulation comparisons at the network outfall and flooding across the network. This multi‐tiered approach not only validates model realism against observed networks but also facilitates understanding of where in the SWMManywhere workflow the synthetic and real UDM diverge, thereby transforming UDM synthesis into an explicit, uncertainty‐driven workflow.
+Because manhole and pipe locations rarely coincide between a synthetic and survey-derived UDM, direct element-by-element comparisons are infeasible [@Chahinian2019-lg]. SWMManywhere therefore implements a comprehensive suite of metrics mapped to each stage of the synthesis pipeline: system description (e.g., pipe lengths, counts), topological structure (e.g., Laplacian and vertex–edge distances, betweenness distributions), hydraulic design (e.g., diameter and depth statistics), and simulation performance (e.g., outfall flows and network flooding volumes). This multi‐tiered approach validates model realism against observed networks and facilitates understanding of where in the SWMManywhere workflow the synthetic and real UDM diverge, thereby transforming UDM synthesis into an explicit, uncertainty‐driven workflow.
 
 ## Outlook
 
