### 2025-04-28 ImperialCollegeLondon/SWMManywhere#364 @cheginit
@barneydobson You need to change `with tempfile.TemporaryDirectory() as temp_dir:` to `with tempfile.TemporaryDirectory(dir=".") as temp_dir:` at https://github.com/ImperialCollegeLondon/SWMManywhere/blob/a0d56448232d28dfd81cfa3ccfdf4cae3986225b/src/swmmanywhere/graphfcns/subcatchment_graphfcns.py#L226

Also, please search for all instances of `tempfile` module use in the codebase and add the `dir` arg to avoid this issue on Windows.
