+++ tests/test_swmmanywhere.py
@@ -40,7 +40,7 @@ def test_run():
 @pytest.mark.parametrize("run", [True, False])
 def test_swmmanywhere(run, wbt_zip_path):
     """Test the swmmanywhere function."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         # Load the config
         test_data_dir = Path(__file__).parent / "test_data"
         defs_dir = Path(__file__).parent.parent / "src" / "swmmanywhere" / "defs"
@@ -122,7 +122,7 @@ def test_swmmanywhere(run, wbt_zip_path):
 
 def test_load_config_file_validation():
     """Test the file validation of the config."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         defs_dir = Path(__file__).parent.parent / "src" / "swmmanywhere" / "defs"
         base_dir = Path(temp_dir)
 
@@ -149,7 +149,7 @@ def test_load_config_file_validation():
 
 def test_load_config_schema_validation():
     """Test the schema validation of the config."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         defs_dir = Path(__file__).parent.parent / "src" / "swmmanywhere" / "defs"
         base_dir = Path(temp_dir)
 
@@ -171,7 +171,7 @@ def test_load_config_schema_validation():
 
 def test_save_config():
     """Test the save_config function."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         temp_dir = Path(temp_dir)
         defs_dir = Path(__file__).parent.parent / "src" / "swmmanywhere" / "defs"
 
@@ -193,7 +193,7 @@ def test_save_config():
 @pytest.mark.downloads
 def test_minimal_req():
     """Test SWMManywhere with minimal info."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         config = {
             "base_dir": Path(temp_dir),
             "project": "my_test",
@@ -205,7 +205,7 @@ def test_minimal_req():
 
 def test_custom_metric():
     """Test adding a custom metric."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         # Load the config
         gf_module = str(Path(__file__).parent / "test_data" / "custom_metrics.py")
 
@@ -238,7 +238,7 @@ def test_custom_metric():
 
 def test_custom_graphfcn():
     """Test adding a custom graphfcn."""
-    with tempfile.TemporaryDirectory() as temp_dir:
+    with tempfile.TemporaryDirectory(dir=".") as temp_dir:
         # Load the config
         gf_module = str(Path(__file__).parent / "test_data" / "custom_graphfcns.py")
 
