+++ docs/notebooks/custom_data_demo.py
@@ -33,7 +33,7 @@ from swmmanywhere.swmmanywhere import swmmanywhere
 from swmmanywhere.utilities import plot_map
 
 # Create temporary directory
-temp_dir = tempfile.TemporaryDirectory()
+temp_dir = tempfile.TemporaryDirectory(dir=".")
 base_dir = Path(temp_dir.name)
 
 # Define minimum viable config
