### 2025-05-05 ImperialCollegeLondon/SWMManywhere#364 @meghnathomas
I was able to get the code to run! Thanks for all your effort in helping me get there!

I was running into issues because the directory in which I installed SWMManywhere was this: "C:\Users\thom017\OneDrive - xxxx\Reviews\JOSS_SWMManywhere"

And the dash in the file path name was (I think) what was causing errors when PySWMM tried to parse the "Rain Gages" section of the INP file:

```
[RAINGAGES]
;;Name           Format    Interval SCF      Source
;;-------------- --------- ------ ------ ----------
1                INTENSITY 00:05    1        FILE       C:\Users\thom017\OneDrive - xxxx\Reviews\JOSS_SWMManywhere\SWMManywhere_again\SWMManywhere\src\swmmanywhere\defs\storm.dat 1          mm
```

After trying this in a different directory, not in the OneDrive, I was able to run everything with no errors.

- [ ] I wanted to add one final comment about the paper: I would recommend concluding the initial "Summary" paragraph with a short description of what the SWMManyhere package brings to the table and how it addresses the gap you laid out.

As soon as you address the final comment and merge PR #377 I will be happy to recommend this paper for acceptance! :-)
