### 2025-05-12 06a5726
Update src/swmmanywhere/post_processing.py
Co-authored-by: Diego Alonso Álvarez <6095790+dalonsoa@users.noreply.github.com>
