+++ src/swmmanywhere/parameters.py
@@ -305,3 +305,13 @@ class MetricEvaluation(BaseModel):
         unit="m",
         description="Scale of the grid for metric evaluation",
     )
+
+    warmup: float = Field(
+        default=0,
+        ge=0,
+        le=1,
+        unit="%/100",
+        description="""Warmup period as a fraction of the total simulation time. This
+            is used to exclude the initial part of the simulation from the metric
+            calculations.""",
+    )
