### 2025-05-15 3b65881
Update swmmanywhere.py
some validation
