### 2025-05-15 0ff23c7
Update parameters.py
terminology
