+++ src/swmmanywhere/metric_utilities.py
@@ -109,7 +109,7 @@ def apply_warmup(results: pd.DataFrame, warmup: float) -> pd.DataFrame:
     Returns:
         pd.DataFrame: The results dataframe with the warmup period applied.
     """
-    return results.iloc[int(warmup * results.shape[0]):]
+    return results.iloc[int(warmup * results.shape[0]) :]
 
 
 def iterate_metrics(
