+++ src/swmmanywhere/parameters.py
@@ -309,7 +309,7 @@ class MetricEvaluation(BaseModel):
     warmup: float = Field(
         default=0,
         ge=0,
-        lt=1,
+        le=1,
         unit="fraction",
         description="""Warmup period as a fraction of the total simulation time. This
             is used to exclude the initial part of the simulation from the metric
