+++ src/swmmanywhere/graphfcns/design_graphfcns.py
@@ -15,6 +15,143 @@ from swmmanywhere.graph_utilities import BaseGraphFunction, register_graphfcn
 from swmmanywhere.logging import logger, verbose
 
 
+def get_designs(hydraulic_design: parameters.HydraulicDesign) -> np.ndarray:
+    """Get the designs for the pipe.
+
+    This function generates a grid of designs for the pipe based on the
+    diameters and depths specified in the hydraulic design parameters. It
+    returns a numpy array of the designs.
+
+    Args:
+        hydraulic_design (parameters.HydraulicDesign): A HydraulicDesign parameter
+            object  
+
+    Returns:
+        designs (np.ndarray): A numpy array of the designs for the pipe
+    """
+    return product(
+        hydraulic_design.diameters,
+        np.linspace(
+            hydraulic_design.min_depth, hydraulic_design.max_depth, 10
+        ),  # TODO should 10 be a param?
+    )
+
+def evaluate_design(
+    ds_elevation: float,
+    chamber_floor: float,
+    edge_length: float,
+    hydraulic_design: parameters.HydraulicDesign,
+    Q: float,
+    diam: float,
+    depth: float,
+    ):
+    """Evaluate the design of a pipe.
+
+    This function evaluates the design of a pipe by calculating the cost,
+    velocity, filling ratio, and feasibility parameters. It returns a dictionary
+    containing the design parameters and their values.
+
+    Args:
+        ds_elevation (float): The downstream elevation
+        chamber_floor (float): The elevation of the chamber floor
+        edge_length (float): The length of the edge
+        hydraulic_design (parameters.HydraulicDesign): A HydraulicDesign parameter
+            object
+        Q (float): The flow rate
+        diam (float): The diameter of the pipe
+        depth (float): The depth of the pipe
+    Returns:
+        dict: A dictionary containing the design parameters and their values
+    """
+    A = np.pi * diam**2 / 4
+    n = 0.012  # mannings n
+    R = A / (np.pi * diam)  # hydraulic radius
+    # TODO... presumably need to check depth > (diam + min_depth)
+
+    elev_diff = chamber_floor - (ds_elevation - depth)
+    slope = elev_diff / edge_length
+    # Always pick a pipe that is feasible without surcharging
+    # if available
+    surcharge_feasibility = 0.0
+    # Use surcharged elevation
+    while slope <= 0:
+        surcharge_feasibility += 0.05
+        slope = (
+            chamber_floor + surcharge_feasibility - (ds_elevation - depth)
+        ) / edge_length
+        # TODO could make the feasibility penalisation increase
+        # when you get above surface_elevation[node]... but
+        # then you'd need a feasibility tracker and an offset
+        # tracker
+    v = (slope**0.5) * (R ** (2 / 3)) / n
+    filling_ratio = Q / (v * A)
+    # buffers from: https://www.polypipe.com/sites/default/files/Specification_Clauses_Underground_Drainage.pdf
+    average_depth = (depth + chamber_floor) / 2
+    V = edge_length * (diam + 0.3) * (average_depth + 0.1)
+    cost = 1.32 / 2000 * (9579.31 * diam**0.5737 + 1163.77 * V**1.31)
+    v_feasibility = max(hydraulic_design.min_v - v, 0) + max(
+        v - hydraulic_design.max_v, 0
+    )
+    fr_feasibility = max(filling_ratio - hydraulic_design.max_fr, 0)
+    """
+    TODO shear stress... got confused here
+    density = 1000
+    dyn_visc = 0.001
+    hydraulic_diameter = 4 * (A * filling_ratio**2) / \
+        (np.pi * diam * filling_ratio)
+    Re = density * v * 2 * (diam / 4) * (filling_ratio ** 2) / dyn_visc
[…]
