### 2025-05-16 c59a5b8
Update design_graphfcns.py
precommit
