### 2025-05-16 77d6556
generalise design params
