### 2025-05-16 35cbc94
Update design_graphfcns.py
move functions to graphfcn
