### 2025-05-16 656919f
Update design_graphfcns.py
doc/typing
