### 2025-05-16 2b35d65
Update design_graphfcns.py
