+++ tests/test_metric_utilities.py
@@ -735,16 +735,25 @@ def test_align_by_shape_and_median_coef(subs, results):
 def test_apply_warmup():
     """Test the apply_warmup function."""
     # Create a DataFrame with a date range
-    df = pd.DataFrame(
-        {
-            "date": pd.date_range(start="2021-01-01", periods=100, freq="s"),
-            "value": range(100),
-        }
+    df = pd.concat(
+        [
+            pd.DataFrame(
+                {
+                    "date": pd.date_range(
+                        start="2021-01-01", periods=100, freq=f"{ds}s"
+                    ),
+                    "value": range(100),
+                    "variable": [variable] * 100,
+                }
+            )
+            for ds, variable in zip([2, 3], ["flow", "flooding"])
+        ],
+        axis=0,
     )
 
     # Apply warmup
     df_warmup = mu.apply_warmup(df, 0.5)
 
     # Check the result
-    assert df_warmup.shape[0] == 50
-    assert df_warmup.date.iloc[0] == pd.to_datetime("2021-01-01 00:00:50")
+    assert df_warmup.shape[0] == 75
+    assert df_warmup.date.min() == pd.to_datetime("2021-01-01 00:02:30")
