### 2025-05-19 bf6a1c4
Update design_graphfcns.py
diego revisions
