### 2025-05-19 e5d7128
Update design_graphfcns.py
