+++ src/swmmanywhere/graphfcns/design_graphfcns.py
@@ -15,290 +15,6 @@ from swmmanywhere.graph_utilities import BaseGraphFunction, register_graphfcn
 from swmmanywhere.logging import logger, verbose
 
 
-def get_designs(hydraulic_design: parameters.HydraulicDesign) -> product:
-    """Get the designs for the pipe.
-
-    This function generates a grid of designs for the pipe based on the
-    diameters and depths specified in the hydraulic design parameters. It
-    returns an iterable product of the designs.
-
-    Args:
-        hydraulic_design (parameters.HydraulicDesign): A HydraulicDesign parameter
-            object
-
-    Returns:
-        product: An iterable product object containing the designs
-    """
-    return product(
-        hydraulic_design.diameters,
-        np.linspace(
-            hydraulic_design.min_depth,
-            hydraulic_design.max_depth,
-            hydraulic_design.depth_nbins,
-        ),
-    )
-
-
-def calculate_cost(V: float, diam: float) -> float:
-    """Calculate the cost of the pipe.
-
-    This function calculates the cost of the pipe based on the excavation volume and
-    diameter.
-
-    Cost equation from: https://doi.org/10.2166/hydro.2016.105
-
-    Args:
-        V (float): The excavation volume of the pipe
-        diam (float): The diameter of the pipe
-    Returns:
-        float: The cost of the pipe in USD
-    """
-    return 1.32 / 2000 * (9579.31 * diam**0.5737 + 1163.77 * V**1.31)
-
-
-def evaluate_design(
-    ds_elevation: float,
-    chamber_floor: float,
-    edge_length: float,
-    hydraulic_design: parameters.HydraulicDesign,
-    Q: float,
-    diam: float,
-    depth: float,
-) -> dict[Hashable, float]:
-    """Evaluate the design of a pipe.
-
-    This function evaluates the design of a pipe by calculating the cost,
-    velocity, filling ratio, and feasibility parameters. It returns a dictionary
-    containing the design parameters and their values.
-
-    Args:
-        ds_elevation (float): The downstream elevation
-        chamber_floor (float): The elevation of the chamber floor
-        edge_length (float): The length of the edge
-        hydraulic_design (parameters.HydraulicDesign): A HydraulicDesign parameter
-            object
-        Q (float): The flow rate
-        diam (float): The diameter of the pipe
-        depth (float): The depth of the pipe
-
-    Returns:
-        dict: A dictionary containing the designed pipe's parameters and values
-    """
-    A = np.pi * diam**2 / 4
-    n = 0.012  # mannings n
-    R = A / (np.pi * diam)  # hydraulic radius
-    # TODO... presumably need to check depth > (diam + min_depth)
-
-    elev_diff = chamber_floor - (ds_elevation - depth)
-    slope = elev_diff / edge_length
-    # Always pick a pipe that is feasible without surcharging
-    # if available
-    surcharge_feasibility = 0.0
-    # Use surcharged elevation
-    while slope <= 0:
-        surcharge_feasibility += 0.05
-        slope = (
-            chamber_floor + surcharge_feasibility - (ds_elevation - depth)
-        ) / edge_length
-        # TODO could make the feasibility penalisation increase
-        # when you get above surface_elevation[node]... but
-        # then you'd need a feasibility tracker and an offset
-        # tracker
-    v = (slope**0.5) * (R ** (2 / 3)) / n
-    filling_ratio = Q / (v * A)
-    # buffers from: https://www.polypipe.com/sites/default/files/Specification_Clauses_Underground_Drainage.pdf
-    average_depth = (depth + chamber_floor) / 2
-    V = edge_length * (diam + 0.3) * (average_depth + 0.1)
[…]
