### 2025-05-22 5984301
Update src/swmmanywhere/graphfcns/design_graphfcns.py
Co-authored-by: Diego Alonso Álvarez <6095790+dalonsoa@users.noreply.github.com>
