### 2025-05-22 2ce8b7f
Update src/swmmanywhere/graphfcns/design_graphfcns.py
Co-authored-by: Diego Alonso Álvarez <6095790+dalonsoa@users.noreply.github.com>
