+++ paper/paper.bib
@@ -1,0 +1,342 @@
+@ARTICLE{Blumensaat2012-hd,
+  title     = "Sewer model development under minimum data requirements",
+  author    = "Blumensaat, Frank and Wolfram, Martin and Krebs, Peter",
+  journal   = "Environ. Earth Sci.",
+  publisher = "Springer Science and Business Media LLC",
+  volume    =  65,
+  number    =  5,
+  pages     = "1427--1437",
+  month     =  mar,
+  year      =  2012,
+  language  = "en",
+  doi = "10.1007/s12665-011-1146-1"
+}
+
+@ARTICLE{Chahinian2019-lg,
+  title     = "Automatic mapping of urban wastewater networks based on manhole
+               cover locations",
+  author    = "Chahinian, Nan{\'e}e and Delenne, Carole and Commandr{\'e},
+               Benjamin and Derras, Mustapha and Deruelle, Laurent and Bailly,
+               Jean-St{\'e}phane",
+  journal   = "Comput. Environ. Urban Syst.",
+  publisher = "Elsevier BV",
+  volume    =  78,
+  number    =  101370,
+  pages     = "101370",
+  month     =  nov,
+  year      =  2019,
+  language  = "en",
+  doi = "10.1016/j.compenvurbsys.2019.101370",
+}
+
+% The entry below contains non-ASCII chars that could not be converted
+% to a LaTeX equivalent.
+@ARTICLE{Dobson2022-wq,
+  title     = "A reduced complexity model with graph partitioning for rapid
+               hydraulic assessment of sewer networks",
+  author    = "Dobson, Barnaby and Watson-Hill, Hannah and Muhandes, Samer and
+               Borup, Morten and Mijic, Ana",
+  abstract  = "AbstractExisting, high‐fidelity models for sewer network
+               modeling are accurate but too slow and inflexible for modern
+               applications such as optimization or scenario analysis. Reduced
+               complexity surrogate modeling has been applied in response to
+               this, however, current approaches are expensive to set up and
+               still require high‐fidelity simulations to derive parameters. In
+               this study, we compare and develop graph partitioning algorithms
+               to automatically group sections of sewer networks into
+               semi‐distributed compartments. These compartments can then be
+               simulated using sewer network information only in the integrated
+               modeling framework, CityWat‐SemiDistributed (CWSD), which has
+               been developed for application to sewer network modeling in this
+               study. We find that combining graph partitioning with CWSD can
+               produce accurate simulations 100--1,000$\times$ faster than
+               existing high‐fidelity modeling. Because we anticipate that many
+               CWSD users will not have high‐fidelity models available, we
+               demonstrate that the approach provides reasonable simulations
+               even under significant parametric uncertainty through a
+               sensitivity analysis. We compare multiple graph partitioning
+               techniques enabling users to specify the spatial aggregation of
+               the partitioned network, also enabling them to preserve key
+               locations for simulation. We test the impact of temporal
+               resolution, finding that accurate simulations can be produced
+               with timesteps up to one hour. Our experiments show a log‐log
+               relationship between temporal/spatial resolution and simulation
+               time, enabling users to pre‐specify the efficiency and accuracy
+               needed for their applications. We expect that the efficiency and
+               flexibility of our approach may facilitate novel applications of
+               sewer network models ranging from continuous simulations for
+               long‐term planning to spatially optimizing the placement of
+               network sensors.",
+  journal   = "Water Resour. Res.",
+  publisher = "American Geophysical Union (AGU)",
+  volume    =  58,
+  number    =  1,
+  month     =  jan,
+  year      =  2022,
[…]
