### 2025-05-24 #7729 @mengqi-z
@cbuahin - Great! Thanks so much for reviewing the paper!

@barneydobson - Now that both reviewers have completed their review, I will proofread the paper to make sure it follows JOSS standards. Once that’s done, I’ll update you on the next steps. Thank you!
