### 2025-05-29 #7729 @mengqi-z
@barneydobson - The paper looks good. We can move forward to the next step.

Could you please complete the following items and let me know once you are done? Thanks!

- [ ] Double check authors and affiliations (including ORCIDs).
- [ ] Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
- [ ] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
- [ ] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
- [ ] Please **post the DOI of the archived version here**.
